Moved componenets and metrics in to their own sub-packages
