+++ taskobra/orm/__init__.py
@@ -1,13 +1,8 @@
 from .base import get_engine, get_session, ORMBase
-from .component import Component
-from .cpu import CPU
-from .gpu import GPU
-from .memory import Memory
-from .metric import Metric
-from .operating_system import OperatingSystem
+from .components import Component, CPU, GPU, Memory, OperatingSystem, Storage
+from .metrics import Metric
 from .role import Role
 from .user import User
 from .snapshot import Snapshot
-from .storage import Storage
 from .system import System
-from .relationships import *
+from .relationships import SystemComponent, user_role_table, UserSystemRole
