+++ taskobra/orm/components/__init__.py
@@ -1,0 +1,6 @@
+from .component import Component
+from .cpu import CPU
+from .gpu import GPU
+from .memory import Memory
+from .operating_system import OperatingSystem
+from .storage import Storage
