+++ taskobra/orm/components/gpu.py
@@ -1,22 +1,22 @@
 # Libraries
 from sqlalchemy import Column, Float, ForeignKey, Integer, String
 # Taskobra
-from taskobra.orm.component import Component
+from taskobra.orm.components import Component
 
 
-class Storage(Component):
-    __tablename__ = "Storage"
+class GPU(Component):
+    __tablename__ = "GPU"
     unique_id = Column(Integer, ForeignKey("Component.unique_id"), primary_key=True)
     manufacturer = Column(String)
     model = Column(String)
-    standard = Column(String)
-    capacity = Column(Float)
-    max_read = Column(Float)
-    max_write = Column(Float)
+    architecture = Column(String)
+    tdp = Column(Integer)
+    core_count = Column(Integer)
+    memory = Column(Float)
 
     __mapper_args__ = {
         "polymorphic_identity": __tablename__,
     }
 
     def __repr__(self):
-        return f"<Storage({self.manufacturer} {self.model})>"
+        return f"<GPU({self.manufacturer} {self.model} ({self.core_count}x{self.architecture}, {self.memory} GB))>"
