+++ taskobra/orm/components/operating_system.py
@@ -1,7 +1,7 @@
 # Libraries
 from sqlalchemy import Column, Float, ForeignKey, Integer, String
 # Taskobra
-from taskobra.orm.component import Component
+from taskobra.orm.components import Component
 
 
 class OperatingSystem(Component):
