+++ taskobra/orm/metrics/__init__.py
@@ -1,0 +1,1 @@
+from .metric import Metric
